Text input for ALL types, bless good abstractions
Removed SliderString (you won this time, Casheww). Also made hooks automagically called, and moved example into its own file.
