+++ ManagedPlacedObjects/ManagedPlacedObjects.csproj
@@ -43,6 +43,9 @@
     <Reference Include="MonoMod.Common">
       <HintPath>..\..\References\MonoMod.Common.dll</HintPath>
     </Reference>
+    <Reference Include="MonoMod.RuntimeDetour">
+      <HintPath>..\..\References\MonoMod.RuntimeDetour.dll</HintPath>
+    </Reference>
     <Reference Include="Partiality">
       <HintPath>..\..\References\Partiality.dll</HintPath>
     </Reference>
@@ -57,6 +60,7 @@
     </Reference>
   </ItemGroup>
   <ItemGroup>
+    <Compile Include="Examples.cs" />
     <Compile Include="ManagedPlacedObjectsMod.cs" />
     <Compile Include="PlacedObjectsManager.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
